Add CNN for calculating R. But there's not val and test in NN.

+++ myscripts/gen_train/gen_train.py
@@ -28,6 +28,8 @@ np.random.seed(seed=config.seed)
 
 d_handler = DataHandler(config)
 
+if(config.mode == "sample_per_d" and os.path.exists(config.SAMPLE_DIR)):
+    exit()
 # Delete Existing smaple file with conforming
 d_handler.delete_sample_files()
